Refactor graph risk analysis in CLI and core modules
