+++ packages/dbt-tools/cli/src/graph-risk-action.ts
@@ -8,6 +8,9 @@ import {
   resolveArtifactPaths,
   shouldOutputJSON,
   validateSafePath,
+  GRAPH_RISK_RANKING_METRICS,
+  isGraphRiskRankingMetric,
+  isGraphRiskResourceType,
   type DbtResourceType,
   type GraphRiskRankingMetric,
 } from "@dbt-tools/core";
@@ -16,12 +19,30 @@ type GraphRiskActionOptions = {
   targetDir?: string;
   runResults?: string;
   top?: number;
-  metric?: GraphRiskRankingMetric;
-  resourceTypes?: DbtResourceType[];
+  /** Raw `--metric` string; validated inside this action. */
+  metric?: string;
+  /** Raw `--resource-types` CSV, or an array (tests / programmatic use). */
+  resourceTypes?: string | DbtResourceType[];
   fields?: string;
   json?: boolean;
   noJson?: boolean;
 };
+
+function parseResourceTypesArg(
+  input: string | DbtResourceType[],
+): { ok: true; value: DbtResourceType[] } | { ok: false; invalid: string[] } {
+  const list = (
+    Array.isArray(input)
+      ? input.map((v) => String(v).trim())
+      : input.split(",").map((s) => s.trim())
+  ).filter((s) => s.length > 0);
+
+  const invalid = list.filter((t) => !isGraphRiskResourceType(t));
+  if (invalid.length > 0) {
+    return { ok: false, invalid };
+  }
+  return { ok: true, value: list as DbtResourceType[] };
+}
 
 export function graphRiskAction(
   manifestPath: string | undefined,
@@ -30,6 +51,35 @@ export function graphRiskAction(
   isTTY: () => boolean,
 ): void {
   try {
+    let metric: GraphRiskRankingMetric | undefined;
+    if (options.metric !== undefined && options.metric !== "") {
+      if (!isGraphRiskRankingMetric(options.metric)) {
+        handleError(
+          new Error(
+            `--metric must be one of: ${GRAPH_RISK_RANKING_METRICS.join(", ")}`,
+          ),
+          isTTY(),
+        );
+        return;
+      }
+      metric = options.metric;
+    }
+
+    let resourceTypes: DbtResourceType[] | undefined;
+    if (options.resourceTypes !== undefined) {
+      const parsed = parseResourceTypesArg(options.resourceTypes);
+      if (!parsed.ok) {
+        handleError(
+          new Error(
+            `--resource-types contains unsupported values: ${parsed.invalid.join(", ")}`,
+          ),
+          isTTY(),
+        );
+        return;
+      }
+      resourceTypes = parsed.value;
+    }
+
     const manifestArtifactPaths = resolveArtifactPaths(
       manifestPath,
       undefined,
@@ -49,23 +99,26 @@ export function graphRiskAction(
       runResults = loadRunResults(runResultsArtifactPaths.runResults);
     }
 
-    const metric = options.metric ?? "overallRiskScore";
+    const resolvedMetric = metric ?? "overallRiskScore";
     const top = options.top ?? 10;
     const analyzer = new GraphRiskAnalyzer({
       manifest,
       ...(runResults ? { runResults } : {}),
       options: {
-        resourceTypes: options.resourceTypes,
+        resourceTypes,
         includeExecution: runResults !== undefined,
         topN: top,
       },
     });
 
     const summary = analyzer.analyze();
-    const topByMetric = analyzer.getTopNodes({ metric, limit: top });
+    const topByMetric = analyzer.getTopNodes({
+      metric: resolvedMetric,
+      limit: top,
+    });
     const report = {
       ...summary,
-      selected_metric: metric,
+      selected_metric: resolvedMetric,
       top_by_metric: topByMetric,
     };
 
@@ -81,7 +134,7 @@ export function graphRiskAction(
     console.log(
       formatGraphRiskReport({
         summary,
-        selectedMetric: metric,
+        selectedMetric: resolvedMetric,
         topByMetric,
       }),
     );
